Fix precision loss issue on decimals

+++ libs/datamodel/connectors/datamodel-connector/src/scalars.rs
@@ -44,7 +44,7 @@ impl ToString for ScalarType {
 pub enum ScalarValue {
     Int(i32),
     Float(f32),
-    Decimal(f32),
+    Decimal(f64),
     Boolean(bool),
     String(String),
     DateTime(DateTime<Utc>),
